requirements.txt
pyyaml
torch
torchvision
opencv-python
numpy
matplotlib
scikit-learn
kaggle
ipywidgets
Pillow